requirements.txt
python-dotenv
streamlit
llama-index
llama-index-vector-stores-qdrant
llama-index-llms-ollama
llama-index-embeddings-ollama
qdrant-client
huggingface_hub
sentence_transformers
docx2txt
python-pptx
llama-index-embeddings-huggingface
llama-index-experimental
arize-phoenix>=4.29.0
#torch
#transformers
#Pillow